auto-heal-function/requirements.txt
azure-functions
azure-identity
azure-mgmt-containerinstance
requests
azure-mgmt-resource
